wake: sketch the approach forr incremental builds

+++ database.h
@@ -40,8 +40,8 @@ struct Database {
     long job,
     const std::string &inputs,   // null separated
     const std::string &outputs); // null separated
-  std::vector<std::string> get_inputs(int job);
-  std::vector<std::string> get_outputs(int job);
+  std::vector<std::string> get_inputs(long job);
+  std::vector<std::string> get_outputs(long job);
 
   void save_output( // call only if needs_build -> true
     long job,
